@@ -4,10 +4,11 @@ import { AppConfigService } from '../../../service/appconfig.service';
 import {DomainService} from '../../../service/domain.service';
 import {UserDataService} from '../../../service/userdata.service';
 import {Component, OnInit, Input, OnDestroy} from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/interval';
+import { Subscription ,  Observable } from 'rxjs';
+
 import { isUndefined, isNullOrUndefined } from 'util';
+import {map} from 'rxjs/operators';
+import {interval} from 'rxjs/internal/observable/interval';
 
 @Component({
   selector: 'nmaas-domain-filter',
@@ -27,7 +28,7 @@ export class DomainFilterComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
       if(this.authService.hasRole('ROLE_SYSTEM_ADMIN')){
-        this.refresh = Observable.interval(10000).subscribe(next => {
+        this.refresh = interval(10000).subscribe(next => {
             if(this.domainService.shouldUpdate()) {
                 this.updateDomains();
                 this.domainService.setUpdateRequiredFlag(false);
@@ -45,7 +46,8 @@ export class DomainFilterComponent implements OnInit, OnDestroy {
     } else {
       this.domains = this.domainService.getMyDomains();
       if(!isUndefined(this.domains) && !this.authService.hasDomainRole(this.appConfig.getNmaasGlobalDomainId(),'ROLE_TOOL_MANAGER') && !this.authService.hasDomainRole(this.appConfig.getNmaasGlobalDomainId(), 'ROLE_OPERATOR')) {
-           this.domains = this.domains.map((domains) => domains.filter((domain) => domain.id !== this.appConfig.getNmaasGlobalDomainId()));
+           this.domains = this.domains.pipe(
+               map((domains) => domains.filter((domain) => domain.id !== this.appConfig.getNmaasGlobalDomainId())));
       }
     }
   }
