@@ -6,13 +6,14 @@ import {UserService} from '../../../service/user.service';
 import {BaseComponent} from '../../common/basecomponent/base.component';
 import {Component, OnInit, Input, Output, EventEmitter, OnChanges, SimpleChanges} from '@angular/core';
 import {AsyncPipe} from '@angular/common';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/shareReplay';
-import 'rxjs/add/operator/take';
+import {Observable, of} from 'rxjs';
+
+
 import { isUndefined } from 'util';
 import {Role, UserRole} from '../../../model/userrole';
 import {UserDataService} from "../../../service/userdata.service";
 import {AuthService} from "../../../auth/auth.service";
+import {map, shareReplay, take} from 'rxjs/operators';
 
 @Component({
   selector: 'nmaas-userslist',
@@ -53,11 +54,13 @@ export class UsersListComponent extends BaseComponent implements OnInit, OnChang
     //console.debug('getDomainName(' + domainId + ')');
     if (this.domainCache.hasData(domainId)) {
       //console.debug('getDomainName(' + domainId + ') from cache');
-      return Observable.of(this.domainCache.getData(domainId).name);
+      return of(this.domainCache.getData(domainId).name);
     } else {
       //console.debug('getDomainName(' + domainId + ') from network');
-      return this.domainService.getOne(domainId).map((domain) => {this.domainCache.setData(domainId, domain); return domain.name})
-              .shareReplay(1).take(1);
+      return this.domainService.getOne(domainId).pipe(
+          map((domain) => {this.domainCache.setData(domainId, domain); return domain.name}),
+          shareReplay(1),
+          take(1));
     }
   }
 
