@@ -1,14 +1,17 @@
+
+import {throwError as observableThrowError, Observable} from 'rxjs';
+import {map} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
 import {AppConfigService} from '../service/appconfig.service';
 import {JwtHelperService} from '@auth0/angular-jwt';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map'
-import 'rxjs/add/operator/timeout';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/throw';
+
+
+
+
 import {isUndefined} from 'util';
 import {Authority} from '../model/authority';
+import {catchError, debounceTime} from 'rxjs/operators';
 
 export class DomainRoles {
   constructor(private domainId: number, private roles: string[] = []) {
@@ -178,9 +181,9 @@ export class AuthService {
   public login(username: string, password: string): Observable<boolean> {
     const headers = new HttpHeaders({'Content-Type': 'application/json', 'Accept': 'application/json'});
     return this.http.post(this.appConfig.config.apiUrl + '/auth/basic/login',
-      JSON.stringify({'username': username, 'password': password}), {headers: headers})
-      .timeout(10000)
-      .map((response: Response) => {
+      JSON.stringify({'username': username, 'password': password}), {headers: headers}).pipe(
+      debounceTime(10000),
+      map((response: Response) => {
         console.debug('Login response: ' + response.statusText);
         // login successful if there's a jwt token in the response
         const token = response && response['token'];
@@ -198,8 +201,8 @@ export class AuthService {
           // return false to indicate failed login
           return false;
         }
-      })
-      .catch((error) => {
+      }),
+      catchError((error) => {
         let message : string;
         if(error.error['message'])
           message = error.error['message'];
@@ -207,8 +210,8 @@ export class AuthService {
           message = 'Server error';
 
         console.debug(error['status']+' - '+message);
-        return Observable.throw(message);
-      });
+        return observableThrowError(message);
+      }));
   }
 
   public propagateSSOLogin(userid: string): Observable<boolean> {
@@ -218,9 +221,9 @@ export class AuthService {
     console.log('propagateSSOLogin ' + userid);
     const headers = new HttpHeaders({'Content-Type': 'application/json', 'Accept': 'application/json'});
     return this.http.post(this.appConfig.config.apiUrl + '/auth/sso/login',
-      JSON.stringify({'userid': userid}), {headers: headers})
-      .timeout(10000)
-      .map((response: Response) => {
+      JSON.stringify({'userid': userid}), {headers: headers}).pipe(
+      debounceTime(10000),
+      map((response: Response) => {
         console.debug('SSO login response: ' + response);
         // login successful if there's a jwt token in the response
         const token = response && response['token'];
@@ -237,8 +240,8 @@ export class AuthService {
           // return false to indicate failed login
           return false;
         }
-      })
-      .catch((error) => {
+      }),
+      catchError((error) => {
         console.debug('SSO login error: ' + error.error['message']);
           let message : string;
           if(error.error['message'])
@@ -246,8 +249,8 @@ export class AuthService {
           else
               message = 'Server error';
 
-          return Observable.throw(message);
-      });
+          return observableThrowError(message);
+      }));
   }
 
   public logout(): void {
