@@ -1,6 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { RegistrationComponent } from './registration.component';
+import {RecaptchaFormsModule} from 'ng-recaptcha/forms';
 
 describe('RegistrationComponent', () => {
   let component: RegistrationComponent;
@@ -8,7 +9,8 @@ describe('RegistrationComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ RegistrationComponent ]
+      declarations: [ RegistrationComponent ],
+      imports: [RecaptchaFormsModule]
     })
     .compileComponents();
   }));
